extracted snapshot method to extension function removed unused listener
